Fix duplicate system name check
Instead of checking for systems with the same (constructor) name, this change checks for the same constructor, i.e. an `instanceof` check. The problem with the previous approach is that a minification system like `terser` would mangle the class name, now with potentially overlapping system names. Which is perfectly valid in a modern modules-based code base, as the name under which the class is imported by the consumer matters, not under which it is defined within the imported module.